+ API end point for Single product

+++ backend-spring-boot/src/main/java/csd/app/product/ProductRepository.java
@@ -1,9 +1,9 @@
 package csd.app.product;
+import java.lang.StackWalker.Option;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-
+public interface ProductRepository extends JpaRepository<Product, Long> {
 }
